Add IPv6 AS Hegemony results
